Fixed a bug in the interrupt routine code
Interrupt routines for @read_char and @read need to place the program
counter back at the beginning of the opcode, but otherwise the program
counter needs to be at the next opcode. Viola was always putting it back
at the beginning of the current opcode, causing Zork Zero to try to
return when there were no other routines on the stack.