Sped up execloop
Moved the checks for the transcription bit and the fixed-pitch bit to
the setbyte and setword routines in memory.py. Instead of checking on
every single instruction to see if these bits have changed, the
interpreter notices when the game tries to change the header byte they
are in.

Similarly, instead of updating the current font size before every
instruction, we wait until the game is trying to access the font size
information in the header, and then update that information before
letting the game see it.